+++ liblfif/include/lfiflib.h
@@ -20,7 +20,7 @@ typedef struct {
   uint64_t       image_width;
   uint64_t       image_height;
   uint64_t       image_count;
-  uint8_t        quality;
+  float          quality;
   uint16_t       max_rgb_value;
   const char    *output_file_name;
 } LFIFCompressStruct;
