+++ tools/include/compress.h
@@ -9,6 +9,6 @@
 #include <cstdint>
 
 void print_usage(const char *argv0);
-bool parse_args(int argc, char *argv[], const char *&input_file_mask, const char *&output_file_name, uint8_t &quality);
+bool parse_args(int argc, char *argv[], const char *&input_file_mask, const char *&output_file_name, float &quality);
 
 #endif
